Remove dataset.code field from GeneralInformation and set it automatically on save
