+++ src/models/DataSetStore.js
@@ -4,7 +4,7 @@ import moment from 'moment';
 import Promise from 'bluebird';
 import { getOwnedPropertyJSON } from 'd2/lib/model/helpers/json';
 import { map, pick, get, filter, flatten, compose, identity, head } from 'lodash/fp';
-import {getCategoryCombos, collectionToArray} from '../utils/Dhis2Helpers';
+import {getCategoryCombos, collectionToArray, getAsyncUniqueValidator} from '../utils/Dhis2Helpers';
 
 // From maintenance-app/src/EditModel/objectActions.js
 const extractErrorMessagesFromResponse = compose(
@@ -16,6 +16,11 @@ const extractErrorMessagesFromResponse = compose(
     map('objectReports'),
     get('typeReports')
 );
+
+const merge = (obj1, obj2) => {
+    _(obj2).each((value, key) => { obj1[key] = value; });
+    return obj1;
+};
 
 export default class DataSetStore {
     constructor(d2, getTranslation) {
@@ -35,6 +40,7 @@ export default class DataSetStore {
         const {associations, dataset} = this.getInitialState();
         this.associations = associations;
         this.dataset = dataset;
+        console.log("store", this);
     }
 
     getInitialModel() {
@@ -101,7 +107,6 @@ export default class DataSetStore {
             const clonedAssociations = _.clone(associations);
             const getOrgUnitIds = (ds) => ds.organisationUnits.toArray().map(ou => ou.id);
             clonedDataset.name = project.name ? project.name : "";
-            clonedDataset.code = project.code ? project.code + " Data Set" : "";
             clonedAssociations.dataInputStartDate =
                 project.startDate ? new Date(project.startDate) : undefined;
             clonedAssociations.dataInputEndDate =
@@ -233,8 +238,24 @@ export default class DataSetStore {
         });
     }
 
+    setCode(dataset) {
+        const {project} = this.associations;
+        const projectCode = project ? project.code : null;
+
+        if (projectCode) {
+            const datasetCode = projectCode + " " + "Data Set";
+            const codeValidator = getAsyncUniqueValidator(this.d2.models.dataSet, "code");
+            return codeValidator(datasetCode)
+                .then(() => merge(dataset.clone(), {code: datasetCode}))
+                .catch(err => dataset);
+        } else {
+            return Promise.resolve(dataset);
+        }
+    }
+
     save() {
         return this.applyDisaggregation(this.dataset)
+            .then(dataset => this.setCode(dataset))
             .then(dataset => this.saveDataset(dataset))
             .then(dataset => {
                 const datasetId = dataset.id;
